final commit

+++ server/excluiComentario.php
@@ -4,15 +4,20 @@ session_start();
 
 $current = $_SESSION['current'];
 
-//Verifica se o usuario é admin ou proprietario do comentário
-
 if (isset($_GET['id'])) {
   $idComment = $_GET['id'];
+  $idLogado = $_SESSION['id'];
 }
 
-$query = "delete from `avalia` where id='$idComment'";
-mysqli_query($connection, $query);
-mysqli_close($connection);
+$query = "select id_usuario from avalia where id=$idComment";
+$result = mysqli_query($connection, $query);
 
+$comment = mysqli_fetch_array($result);
+$idUsuario = $comment['id_usuario'];
+if ($idUsuario == $idLogado || $idLogado == 1) {
+  $query = "delete from `avalia` where id='$idComment'";
+  mysqli_query($connection, $query);
+  mysqli_close($connection);
+}
 header('Location: ' . $current);
 ?>
